Create Mainscript.py
The first code complying with the following:
-Downloading a Vid+Aud < 720p
-Downloading a Vid+Aud == 720p
-Downloading a Vid+Aud > 720p
--Downloading only Aud obtaining an .mp3 file

The topics mentioned were the goals of the program when i got the idea. Im sure it could be not only better in usage and visualization, it could surely be written in a more pythonic and elegant manner. Keeping in mind its my first serious program i wont complain, lets see what the future awaits!